Implemented login views and user models

+++ views/vets.py
@@ -1,14 +1,16 @@
+from flask_login import login_required, current_user
 from flask_restful import reqparse, Resource
-from flask import redirect
+from flask import redirect, abort, session
 from models.vets import Vets
 
 
 class VetsAPI(Resource):
+    method_decorators = [login_required]
 
     def __init__(self):
         self.parser = reqparse.RequestParser()
         self.parser.add_argument('name', type=str)
-        self.parser.add_argument('age', type=int)
+        self.parser.add_argument('age', type=str)
 
     def get(self, id):
         u = Vets.get(id=id)
@@ -17,22 +19,29 @@ class VetsAPI(Resource):
         return {}, 404
 
     def put(self, id):
-        args = self.parser.parse_args()
-        u = Vets.get(id=id)
-        if u and args:
-            u.update(**args)
-            return u.__dict__
+        if current_user.is_admin:
+            args = self.parser.parse_args()
+            u = Vets.get(id=id)
+            if u and args:
+                u.update(**args)
+                return u.__dict__
+        else:
+            abort(403)
         return {}, 404
 
     def delete(self, id):
-        u = Vets.get(id=id)
-        if u:
-            r = u.__dict__
-            u.delete()
-            return r, 200
+        if current_user.is_admin:
+            u = Vets.get(id=id)
+            if u:
+                r = u.__dict__
+                u.delete()
+                return r, 200
+        else:
+            abort(403)
 
 
 class VetsListAPI(Resource):
+    method_decorators = [login_required]
 
     def __init__(self):
         self.parser = reqparse.RequestParser()
@@ -40,16 +49,21 @@ class VetsListAPI(Resource):
         self.parser.add_argument('age', type=str)
 
     def get(self):
-        qs = Vets.filter()
-        if qs:
-            r = [u.__dict__ for u in qs]
-            return r
-        return {}, 404
+        if current_user.is_admin:
+            print(session)
+            qs = Vets.filter()
+            if qs:
+                r = [u.__dict__ for u in qs]
+                return r
+        else:
+            abort(403)
+        abort(404)
 
     def post(self):
         args = self.parser.parse_args()
         if args:
             u = Vets.create(**args)
             if u:
-                return redirect('/vets/')
+                vet = Vets.filter(**args).__getitem__(0)
+                return redirect('/vets/{}'.format(vet.id))
         return {}, 404
